fix exit

+++ gh_exe.ps1
@@ -33,6 +33,9 @@ param (
 
 # absolute path is required, because current directory is not the same as the script directory
 $file_name, $etag = ~/.gpm/scripts/lib/gh_dl.ps1 -url $url -ScriptBlock { $_.name -match $pattern }
+if ($LASTEXITCODE -ne 0) {
+	exit $LASTEXITCODE
+}
 
 7z e $file_name "-o$dest" -r $name -y -bso0 -bsp0 && Remove-Item $file_name
